Show percentage transfer progress

+++ p2p-lib/src/main/java/org/smartregister/p2p/search/ui/P2PReceiverViewModel.kt
@@ -244,7 +244,7 @@ class P2PReceiverViewModel(
           TransferProgress(
             totalRecordCount = totalRecords,
             transferredRecordCount = totalReceivedRecords,
-            percentageTransferred = percentageReceived.toFloat()
+            percentageTransferred = percentageReceived
           )
         )
       }
