+++ p2p-lib/src/main/java/org/smartregister/p2p/search/ui/P2PSenderViewModel.kt
@@ -234,7 +234,7 @@ class P2PSenderViewModel(
           TransferProgress(
             totalRecordCount = totalRecords,
             transferredRecordCount = totalSentRecords,
-            percentageTransferred = percentageSent.toFloat()
+            percentageTransferred = percentageSent
           )
         )
       }
